# perf: document fast init