--- a/jina/types/document/__init__.py
+++ b/jina/types/document/__init__.py
@@ -229,14 +229,15 @@ class Document(ProtoTypeMixin, VersionedMixin):
         """
         Document.CNT += 1
 
-        self._pb_body = jina_pb2.DocumentProto()
         try:
             if isinstance(document, jina_pb2.DocumentProto):
                 if copy:
+                    self._pb_body = jina_pb2.DocumentProto()
                     self._pb_body.CopyFrom(document)
                 else:
                     self._pb_body = document
             elif isinstance(document, bytes):
+                self._pb_body = jina_pb2.DocumentProto()
                 self._pb_body.ParseFromString(document)
             elif isinstance(document, (dict, str)):
                 if isinstance(document, str):
@@ -270,6 +271,7 @@ class Document(ProtoTypeMixin, VersionedMixin):
                     )
                 )
 
+                self._pb_body = jina_pb2.DocumentProto()
                 if support_fields.issuperset(user_fields):
                     json_format.ParseDict(document, self._pb_body)
                 else:
@@ -297,12 +299,16 @@ class Document(ProtoTypeMixin, VersionedMixin):
                             )
             elif isinstance(document, Document):
                 if copy:
+                    self._pb_body = jina_pb2.DocumentProto()
                     self._pb_body.CopyFrom(document.proto)
                 else:
                     self._pb_body = document.proto
             elif document is not None:
                 # note ``None`` is not considered as a bad type
                 raise ValueError(f'{typename(document)} is not recognizable')
+            else:
+                # create an empty document
+                self._pb_body = jina_pb2.DocumentProto()
         except Exception as ex:
             raise BadDocType(
                 f'fail to construct a document from {document}, '