This speeds up DocumentArray.__iter__ heavily since Document constructor becomes much faster
Consequently, all methods that use __iter__ become faster.
When developing this, we had to set mermaid_id efficiently. A similar issue was raised before: https://github.com/jina-ai/jina/issues/2582

To solve it, we used a class attribute counter (to make mermaid_id unique based on document id which is already unique)

results:
|    |      master      |  perf-document-fast-init ( ff87e72 ) | perf-document-fast-init (ec5961d) | perf-document-fast-init (latest) |
|----------|:-------------:|------:|------:|------:|
| document array iter |  1.16 s | 0.32 s | 0.50 s | 0.36 s |
| document array get_attributes('text') |    1.32 s   |   0.42 s | 0.61 s | 0.43 s |
| document array get_attributes('text', 'embedding') | 3.32 s | 2.06 s | 2.21 s | 2.17 s |
| document construct by pb_body | 1.3 s | 0.68 s | 0.6 s | 0.45 s |