@davidbp 2021-09-03
> To solve it, we used a class attribute counter (to make mermaid_id unique based on document id which is already unique)

I want to leave a comment here because the solution might seem unjustified/weird.
While we were doing this with Alaeddine we found out that creating the unique ids took a lot of time (to me this was not exected at least). You pay a  big price  for iterating  and calling the previous unique id mechanism. 

Just as a reference:
```
In [7]: %timeit [uuid.uuid4() for i in range(100_000)]
318 ms ± 1.78 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [8]: %timeit [i for i in range(100_000)]
2.87 ms ± 27.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```