@hanxiao 2021-09-06
i moved the Proto constructor inside, this should give a small performance boost and should result in similar performance as your first version. Could you give it a check?