@alaeddine-13 2021-09-07
thanks for your help @hanxiao 
I updated the table, now back to the first version performance
Also, after the refactor commit, also DAM got a boost:

- dam iter:
     * perf-document-fast-init : 2.32 s
     * master : 3.17 s

- dam get_attributes('text'):
     * perf-document-fast-init : 2.03 s
     * master : 3.36 s